backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.10.0
# AI dependencies (AI Engineer will use these)
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.0
Sastrawi>=1.0.0
scipy>=1.14.0
